refactor(markdown-utils): move mermaid renderer into the shared package with idPrefix param (1.2.0)
Final markdown/image host↔plugin dedup. Moves mermaidRender.ts (DOM scan + lazy
mermaid-runtime load + SVG adoption) and mermaidExtension.ts (comment-only drift)
into @mulmoclaude/markdown-utils. The only host↔plugin functional difference was
the per-diagram DOM id prefix; parameterized as
renderMermaidNodes(root, labels?, idPrefix = "mulmo-mermaid"), threaded through
renderOne → nextRenderId. Added `mermaid` to the package peer/dev deps.

useMermaid.ts stays per-side on purpose: the host wires vue-i18n, the plugin wires
its own useT — genuinely environment-specific. The plugin's composable passes
"mulmo-mermaid-plugin" to keep its prefix; the host uses the default. So both sides
render exactly as before (the prefix only feeds mermaid.render's invisible SVG root
id). Redundant test/plugins/markdown/test_mermaidRender.ts deleted (near-dup of the
host test; both only exercised adoptSvg).

Clean MINOR (1.2.0): no consumer sweep (^1.x accepts it); launcher + markdown-plugin
ranges bumped ^1.2.0 for hygiene. jscpd (spreadsheet excl): 2163 -> 2036 lines
(1.80% -> 1.69%). build + typecheck + lint (0 new) green; mermaid unit tests 15/15.
App-level /verify is user-triggered (agent could not run it) — recommended before
merge; CI e2e also covers the rendering path. Release prereq: publish 1.2.0 first.
