+++ packages/plugins/markdown-plugin/package.json
@@ -36,7 +36,7 @@
   },
   "dependencies": {
     "@marp-team/marp-core": "^4.4.0",
-    "@mulmoclaude/markdown-utils": "^1.1.0",
+    "@mulmoclaude/markdown-utils": "^1.2.0",
     "js-yaml": "^5.2.1",
     "marked": "^18.0.6",
     "mermaid": "^11.16.0"
