+++ packages/plugins/markdown-plugin/src/plugins/markdown/View.vue
@@ -154,7 +154,7 @@ import type { ToolResult } from "gui-chat-protocol";
 import { isFilePath, type MarkdownToolData } from "./definition";
 import { rewriteMarkdownImageRefs } from "@mulmoclaude/markdown-utils/image/rewriteMarkdownImageRefs";
 import { findTaskLines, makeTasksInteractive, toggleTaskAt } from "@mulmoclaude/markdown-utils/markdown/taskList";
-import { mermaidExtension } from "../../utils/markdown/mermaidExtension";
+import { mermaidExtension } from "@mulmoclaude/markdown-utils/markdown/mermaidExtension";
 import { useMermaidRenderer } from "../../utils/markdown/useMermaid";
 import { usePdfExport } from "./usePdfExport";
 import { handleExternalLinkClick } from "@mulmoclaude/markdown-utils/dom/externalLink";
