+++ packages/plugins/markdown-plugin/src/utils/markdown/useMermaid.ts
@@ -3,7 +3,7 @@
 // bundle. Behavior is identical.
 
 import { onMounted, watch, nextTick, type Ref } from "vue";
-import { renderMermaidNodes, type MermaidRenderLabels } from "./mermaidRender";
+import { renderMermaidNodes, type MermaidRenderLabels } from "@mulmoclaude/markdown-utils/markdown/mermaidRender";
 import { useT } from "../../lang";
 
 export function useMermaidRenderer(containerRef: Ref<HTMLElement | null | undefined>, sourceRef: Ref<unknown>): void {
@@ -17,7 +17,7 @@ export function useMermaidRenderer(containerRef: Ref<HTMLElement | null | undefi
   };
   const run = async (): Promise<void> => {
     await nextTick();
-    await renderMermaidNodes(containerRef.value ?? null, labels);
+    await renderMermaidNodes(containerRef.value ?? null, labels, "mulmo-mermaid-plugin");
   };
   onMounted(() => {
     void run();
